feat: Further simplified math
I was originally doing rotations mod 100 as an intermediate step to make sure that it's always in interval [0, 99] but this isn't necessary since (x + r + 100k) mod 100 = (x + (r + 100k) mod 100) mod 100.

This is because in the first equation you get (x + r + 100k) mod 100. We can eliminate 100k since it won't change so this is (x + r) mod 100. In the second equation we get (x + r)
mod 100 since r + 100k mod 100 is just r